pyproject.toml
[project]
name = "ucl-solkoff"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "duckdb>=1.4.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
